+++ lib/l10n/app_localizations_en.dart
@@ -1012,6 +1012,20 @@ class AppLocalizationsEn extends AppLocalizations {
       'Bleeding edge features. Expect bugs and instability.';
 
   @override
+  String whatsNewWelcome(String version) {
+    return 'Welcome to TKit $version!';
+  }
+
+  @override
+  String get whatsNewSubtitle => 'Here\'s what\'s new in this update';
+
+  @override
+  String get whatsNewTag => 'NEW';
+
+  @override
+  String get whatsNewGotIt => 'Got it, thanks!';
+
+  @override
   String get fallbackBehaviorDoNothing => 'Do Nothing';
 
   @override